+++ src/components/CopyText.tsx
