+++ src/proroute/prodrawer/ProDrawerNotAdmin.tsx
@@ -20,14 +20,14 @@ import Info from '@mui/icons-material/Info'
 import Dashboard from '@mui/icons-material/Dashboard'
 import LibraryBooks from '@mui/icons-material/LibraryBooks'
 import Person from '@mui/icons-material/Person'
-import { PathBreadcrumbs } from '../proutils/PathBreadcrumbs'
+import { PathBreadcrumbs } from '~components/PathBreadcrumbs'
 import { useTranslation } from 'react-i18next'
 import { ListItemButton, Switch } from '@mui/material'
 import { useDispatch, useSelector } from 'react-redux'
 import { useLocation, useRouteMatch, Link } from 'react-router-dom'
 import { authSelectors, serverSelectors } from '~store/types'
 import { logout } from '~store/modules/auth/actions'
-import { NmLink } from '../proutils'
+import { NmLink } from '~components/Link'
 import { UI_VERSION } from 'src/config'
 import Logo from '../../../netmaker-logo.png'
 import DarkLogo from '../../../netmaker-logo-2.png'
